+++ src/test/java/org/broadinstitute/hellbender/tools/exome/orientationbiasvariantfilter/OrientationBiasFiltererUnitTest.java
@@ -6,6 +6,7 @@ import htsjdk.variant.vcf.VCFConstants;
 import org.apache.commons.lang3.tuple.Pair;
 import org.broadinstitute.hellbender.engine.FeatureDataSource;
 import org.broadinstitute.hellbender.tools.picard.analysis.artifacts.Transition;
+import org.broadinstitute.hellbender.utils.UniqueIDWrapper;
 import org.broadinstitute.hellbender.utils.test.BaseTest;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
@@ -155,7 +156,7 @@ public class OrientationBiasFiltererUnitTest extends BaseTest {
         final List<String> sampleNames = updatedVariants.get(0).getSampleNamesOrderedByName();
 
         // Create a mapping from sample name to a genotype->variant context map
-        final Map<String, SortedMap<OrientationBiasFilterer.UniqueIDWrapper<Genotype>, VariantContext>> sampleNameToVariants = OrientationBiasFilterer.createSampleToGenotypeVariantContextSortedMap(sampleNames, updatedVariants);
+        final Map<String, SortedMap<UniqueIDWrapper<Genotype>, VariantContext>> sampleNameToVariants = OrientationBiasFilterer.createSampleToGenotypeVariantContextSortedMap(sampleNames, updatedVariants);
     }
 
 
@@ -194,7 +195,7 @@ public class OrientationBiasFiltererUnitTest extends BaseTest {
 
         // Do the test
         // Create a mapping from sample name to a genotype->variant context map with the second map sorted by p_artifact (i.e. OrientationBiasFilterConstants.P_ARTIFACT_FIELD_NAME)
-        final Map<String, SortedMap<OrientationBiasFilterer.UniqueIDWrapper<Genotype>, VariantContext>> sampleNameToVariants = OrientationBiasFilterer.createSampleToGenotypeVariantContextSortedMap(sampleNames, updatedVariants);
+        final Map<String, SortedMap<UniqueIDWrapper<Genotype>, VariantContext>> sampleNameToVariants = OrientationBiasFilterer.createSampleToGenotypeVariantContextSortedMap(sampleNames, updatedVariants);
         Assert.assertEquals(sampleNameToVariants.keySet().size(), 2);
         Assert.assertTrue(sampleNameToVariants.keySet().contains("TUMOR"));
         Assert.assertTrue(sampleNameToVariants.keySet().contains("NORMAL"));
@@ -206,7 +207,7 @@ public class OrientationBiasFiltererUnitTest extends BaseTest {
 
         // Check that the sorted map is getting smaller (or same) values of p_artifact and not staying put.
         double previousPArtifact = Double.POSITIVE_INFINITY;
-        for (final OrientationBiasFilterer.UniqueIDWrapper<Genotype> genotypeTumorWrapped : sampleNameToVariants.get("TUMOR").keySet()) {
+        for (final UniqueIDWrapper<Genotype> genotypeTumorWrapped : sampleNameToVariants.get("TUMOR").keySet()) {
             final Genotype genotypeTumor = genotypeTumorWrapped.getWrapped();
             final Double pArtifact = OrientationBiasUtils.getGenotypeDouble(genotypeTumor, OrientationBiasFilterConstants.P_ARTIFACT_FIELD_NAME, Double.POSITIVE_INFINITY);
             Assert.assertNotNull(pArtifact);
